pyproject.toml
[tool.poetry]
name = "osi"
version = "0.1.0"
description = ""
authors = ["Ana Paula Gomes <apgomes88@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
osi = "osi.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.8"
jsonschema = "^4.23.0"
requests = "^2.32.3"
pygments = "^2.19.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
